[BE] REFACTOR: 논의했던 파사드 패턴으로 적용
