+++ backend/src/main/java/org/ftclub/cabinet/lent/service/LentExceptionHandlerService.java
@@ -8,9 +8,6 @@ import org.ftclub.cabinet.exception.ServiceException;
 import org.ftclub.cabinet.lent.domain.LentHistory;
 import org.ftclub.cabinet.lent.domain.LentPolicyStatus;
 import org.ftclub.cabinet.lent.repository.LentRepository;
-import org.ftclub.cabinet.user.domain.User;
-import org.ftclub.cabinet.user.domain.UserRole;
-import org.ftclub.cabinet.user.repository.UserRepository;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -18,21 +15,7 @@ import org.springframework.stereotype.Service;
 public class LentExceptionHandlerService {
 
 	private final LentRepository lentRepository;
-	private final UserRepository userRepository;
 	private final CabinetExceptionHandlerService cabinetExceptionHandler;
-
-	public User getUser(Long userId) {
-		return userRepository.findById(userId)
-				.orElseThrow(() -> new ServiceException(ExceptionStatus.NOT_FOUND_USER));
-	}
-
-	public User getClubUser(Long userId) {
-		User user = getUser(userId);
-		if (!user.isUserRole(UserRole.CLUB)) {
-			throw new ServiceException(ExceptionStatus.NOT_FOUND_USER);
-		}
-		return user;
-	}
 
 	public LentHistory getActiveLentHistoryWithUserIdAndCabinetId(Long userId, Long cabinetId) {
 		LentHistory ret = getActiveLentHistoryWithUserId(userId);
