+++ backend/src/main/java/org/ftclub/cabinet/lent/service/LentService.java
@@ -1,9 +1,5 @@
 package org.ftclub.cabinet.lent.service;
 
-import java.util.List;
-import org.ftclub.cabinet.dto.LentDto;
-import org.ftclub.cabinet.dto.LentHistoryPaginationDto;
-
 public interface LentService {
 
 	void startLentCabinet(Long userId, Long cabinetId);
@@ -13,11 +9,4 @@ public interface LentService {
 	void endLentCabinet(Long userId);
 
 	void terminateLentCabinet(Long userId);
-
-	LentHistoryPaginationDto getAllUserLentHistories(Long userId, Integer page, Integer length);
-
-	LentHistoryPaginationDto getAllCabinetLentHistories(Long cabinetId, Integer page,
-			Integer length);
-
-	List<LentDto> getLentDtoList(Long cabinetId);
 }
